refactor :#11-구글-지도-장소-검색-api-구현 - searchNearBy, searchText 에서 공통으로 사용되는 Circle, Center 클래스 파일로 따로 빼서 작성 - 구글 맵 관련 dto를 관리의 편의성을 위해 google-map 패키지로 이동
